restructure, rework and container
This commit starts the entry point for users of this library in the file lib.zig.

It restuctures the code directory into subdirectories for each sub-part for a better overview.

It implements the Container struct and the list functions as with Image. Image is also rewritten to use an Arena Allocator internally like Container.

Signed-off-by: Kilian Hanich <khanich.opensource@gmx.de>