create: made nexpodd_path an argument instead of envvar
This removes the need to add this to your environment if you need to
specify a custom path and makes it more explicit to users.

At the same time this makes testing easier since you don't need to
export it into your environment. The system tests expect it as their
first argument.

Signed-off-by: Kilian Hanich <khanich.opensource@gmx.de>