release 0.1.0
This marks the first release (and the MVP) of this project.

It should be enough to use as a minimal base for other projects,
but more features will be added over time.

Additionally done with this release:

updated zeit dependency

updated arch image recipe to use the commit of the shim package
right before the release

Signed-off-by: Kilian Hanich <khanich.opensource@gmx.de>